like jwt

+++ internal/users/repository.go
@@ -3,8 +3,8 @@ package users
 import "github.com/Lemuriets/diary/model"
 
 type Repository interface {
-	GetById(id uint64) (model.User, error)
-	Update(id uint64, updateFields map[string]interface{}) error
-	Delete(id uint64) error
-	MultipleDelete(ids []uint64) error
+	GetById(id uint) (model.User, error)
+	Update(id uint, updateFields map[string]interface{}) error
+	Delete(id uint) error
+	MultipleDelete(ids []uint) error
 }
