+++ internal/users/repository/repository.go
@@ -15,7 +15,7 @@ func NewRepository(db *gorm.DB) *Repository {
 	}
 }
 
-func (repo *Repository) GetById(id uint64) (model.User, error) {
+func (repo *Repository) GetById(id uint) (model.User, error) {
 	var user model.User
 
 	result := repo.DB.Where("ID = ?", id).First(&user)
@@ -23,19 +23,19 @@ func (repo *Repository) GetById(id uint64) (model.User, error) {
 	return user, result.Error
 }
 
-func (repo *Repository) Update(id uint64, updateFields map[string]interface{}) error {
+func (repo *Repository) Update(id uint, updateFields map[string]interface{}) error {
 	result := repo.DB.Model(&model.User{}).Where("id = ?", id).Updates(updateFields)
 
 	return result.Error
 }
 
-func (repo *Repository) Delete(id uint64) error {
+func (repo *Repository) Delete(id uint) error {
 	result := repo.DB.Unscoped().Delete(&model.User{}, id)
 
 	return result.Error
 }
 
-func (repo *Repository) MultipleDelete(ids []uint64) error {
+func (repo *Repository) MultipleDelete(ids []uint) error {
 	result := repo.DB.Unscoped().Delete(&model.User{}, ids)
 
 	return result.Error
