+++ internal/users/usecase.go
@@ -3,8 +3,8 @@ package users
 import "github.com/Lemuriets/diary/model"
 
 type UseCase interface {
-	GetById(id uint64) model.User
-	Update(id uint64, updateFields map[string]interface{}) error
-	Delete(id uint64) error
+	GetById(id uint) model.User
+	Update(id uint, updateFields map[string]interface{}) error
+	Delete(id uint) error
 	AddClass(user *model.User, class model.Class) error
 }
